requirements.txt
streamlit
pillow
opencv-python
tensorflow
tensorflow_hub
matplotlib
pandas
requests
numpy
python-docx
pytesseract
replicate